Created a new renderer to draw glyphs using Processing's built-in drawing methods. This renderer is called PGraphicsTextPageRenderer and replaces the older Java2DTextPageRenderer (which is broken in this version).
The new renderer allows NextText to be used with any Processing graphics contexts, such as OpenGL and P3D. It also allows the use of recorders, such as PDF and MovieMaker.

The new renderer also uses PFonts as its base (the actual Fonts are embedded in the PFont objects). This means faster drawing for non-deformed, non-stroked glyphs. It also means that TextObjectGlyph's font object had to be updated.

For deformed and stroked glyphs, the glyph shapes are tesselated using the OpenGL GLU library included with the Processing distribution. This  appears to be pretty fast but more testing should be done.

Finally, instead of Book mimicking Processing's textFont(...), textAlign(...) methods, it now reads the values directly from the PApplet and applies them to any text it builds.